Move tiny pipeline tests to integration
