+++ tests/integration/support/__init__.py
@@ -1,0 +1,1 @@
+"""Shared support helpers for fixture-scale integration tests."""
