+++ tests/integration/support/pipeline_workspace.py
